requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.20
python-docx==1.1.2
reportlab==4.2.5
